continue developing version

+++ src/expressions.hpp
@@ -13,10 +13,10 @@ namespace pil {
 class Expressions {
     public:
         const Reference *add(const std::string &name, const Reference &value);
-        const Reference *get(uint64_t id) { return values + id; };
+        const Reference *get(uint id) { return values + id; };
         void map(void *data) { evaluations = (FrElement *)data; }
-        FrElement getEvaluation(uint64_t id, uint64_t w) { return evaluations[ w * count + id]; }
-        std::string getName(uint64_t id) { return values[id].name; }
+        FrElement getEvaluation(uint id, uint w) { return evaluations[ w * count + id]; }
+        std::string getName(uint id) { return values[id].name; }
     protected:
         int count = 0;
         int size = 0;
